requirements.txt
numpy==1.24.1
Pillow==9.3.0
torch==2.1.0+cu118
torchvision==0.16.0+cu118
pytorch-metric-learning==2.3.0
tqdm==4.66.1
omegaconf==2.3.0
wandb==0.16.0
